refactor(web): extract CertTable, CertMobileList, and PaginationBar (#52)
* fix(web): ignore stale certificate detail responses

Use request generation tokens so rapid cert switches and closes cannot apply outdated details, issuer data, or errors.



* fix(web): drive Sonner theme from createThemeStore

Pass the app theme store into the toaster and drop unused mode-watcher so toast chrome matches light/dark toggle.



* refactor(web): extract CertTable, CertMobileList, and PaginationBar

Split the desktop table, mobile card list, and pagination chrome out of App.svelte so inventory presentation can evolve without touching load/URL orchestration.



---------
